requirements.txt
aiohttp
# for txt2img_simple
requests
# for img2img
sd-parsers
# for img2img and interrogate
Pillow